Move to BaseActivator
