+++ core/src/main/java/io/stargate/core/activator/BaseActivator.java
@@ -17,7 +17,12 @@ package io.stargate.core.activator;
 
 import com.codahale.metrics.health.HealthCheck;
 import com.codahale.metrics.health.HealthCheckRegistry;
-import java.util.*;
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Hashtable;
 import java.util.List;
 import java.util.Objects;
@@ -36,6 +41,8 @@ import org.slf4j.LoggerFactory;
 
 public abstract class BaseActivator implements BundleActivator {
   private static final Logger logger = LoggerFactory.getLogger(BaseActivator.class);
+
+  private String BASEDIR = System.getProperty("stargate.basedir");
 
   private final ServicePointer<HealthCheckRegistry> healthCheckRegistry =
       ServicePointer.create(HealthCheckRegistry.class);
@@ -258,8 +265,20 @@ public abstract class BaseActivator implements BundleActivator {
     // no-op
   }
 
-  protected String getActivatorName() {
-    return activatorName;
+  /**
+   * Creates a throw away data directory for Stargate's ephemeral files. It can be overridden using
+   * the system property {@code "stargate.basedir"} which can be useful on systems where the
+   * temporary directory is periodically cleaned.
+   *
+   * @return A file handle to the base directory.
+   * @throws IOException if the base directory is invalid or unable to be created.
+   */
+  protected File getBaseDir() throws IOException {
+    if (BASEDIR == null || BASEDIR.isEmpty()) {
+      return Files.createTempDirectory("stargate-" + activatorName).toFile();
+    } else {
+      return Files.createDirectories(Paths.get(BASEDIR)).toFile();
+    }
   }
 
   /**
