+++ persistence-dse-6.8/src/main/java/io/stargate/db/dse/DsePersistenceActivator.java
@@ -5,9 +5,9 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableList.Builder;
 import io.stargate.auth.AuthorizationProcessor;
 import io.stargate.auth.AuthorizationService;
+import io.stargate.core.activator.BaseActivator;
 import io.stargate.core.metrics.api.Metrics;
 import io.stargate.db.Persistence;
-import io.stargate.db.PersistenceActivator;
 import io.stargate.db.datastore.common.StargateConfigSnitch;
 import io.stargate.db.datastore.common.StargateSeedProvider;
 import io.stargate.db.dse.impl.DelegatingAuthorizer;
@@ -35,7 +35,7 @@ import org.apache.cassandra.metrics.CassandraMetricsRegistry;
 import org.apache.cassandra.utils.FBUtilities;
 import org.apache.commons.io.FileUtils;
 
-public class DsePersistenceActivator extends PersistenceActivator {
+public class DsePersistenceActivator extends BaseActivator {
 
   private static final String AUTHZ_PROCESSOR_ID =
       System.getProperty("stargate.authorization.processor.id");
